refactor(test): consolidate shortcut-controller tests into single integration file
Merge unit tests (42) and integration tests (7) into one integration file
using bidirectional mock pattern with unsubscribe tracking. Add missing
ArrowLeft/ArrowRight key normalization coverage. Reorganize into behavioral
groups: view registration, Alt+X activation, Alt release, preventDefault
regression, shortcut key emission, edge cases, cleanup.
